Fix: wrong credentials login bug.
Update: reorganized some code.
